Move python host declaration to top to fix provider initialization
Python plugins were having a frustrating issue where they would fail
with a missing python provider error, but _only_ if I started up vim
with a python file directly. If I opened vim, then opened a python file,
the plugins would work as expected.

This appears to be due to the order of when the `python3_host-prog`
variable was set. Moving it to the top of the initialization properly
sets the host in time, and now the plugins load without issue.